feat: abstract ui into its own class
